+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.8.0] - 2025-07-18
+
 ### Added
 
 - Add `--allow-staged`
@@ -99,7 +101,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-07-09
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.8.0
 [0.7.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.7.0
 [0.6.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.6.0
 [0.5.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.5.0